fixed elvis_meta_SUITE:dialyzer/1 test
The dialyzer test fails because test/example/pass_no_useless_seqbind.erl
does not compile. This is because the seqbind compile transorm is
incorrectly defined for the bindings in the file. However, this is
intentional as we want to assert that elvis does not warn
about the use of the parse transform. As the binding are not really
required for the purposes of this test, i have removed the seqbind
bindings to keep the dialyzer test happy.